+++ convtest.cc
@@ -7,14 +7,14 @@ using std::endl;
 
 int main()
 {	
-  unsigned int i;
-  unsigned int n=32;
-  unsigned int np=n/2+1;
+  size_t i;
+  size_t n=32;
+  size_t np=n/2+1;
   
   Complex f[np],g[np],h[np];
   Complex d[]={-5,Complex(3,1),Complex(4,-2),Complex(-3,1),Complex(0,-2),Complex(0,1),Complex(4,0),Complex(-3,-1),Complex(1,2),Complex(2,1),Complex(3,1)};
 	
-  unsigned int m=sizeof(d)/sizeof(Complex);
+  size_t m=sizeof(d)/sizeof(Complex);
   convolution convolve(n,m);
 
   for(i=0; i < m; i++) f[i]=d[i];
